Fixed up layout
Getting images to fill the grid cells is mostly about adding width: 100% to the image element style.
Also added an extra line
 / 1fr  1fr  1fr  1fr  1fr  1fr  1fr  1fr  1fr  1fr  1fr  1fr
at the end of the grid-template, which forces the layout grid to have equal width columns (sometimes necessary if the images being laid out in the grid have different sizes).
Added in the definitions of the fonts in the design file